Use "puppeteer-core" instead of full "puppeteer" NPM package
The full "puppeteer" NPM package also takes care of downloading
and managing the web browsers. But as we use the system browser
we do not need this functionality.

Installing just the "puppeteer-core" package makes the NPM bundle
a little bit smaller.

See more details in https://pptr.dev/guides/installation